+++ vss_cli/plugins/compute_plugins/iso.py
@@ -35,7 +35,7 @@ def compute_iso_public_ls(ctx: Configuration, filter_by, show_all, sort, page):
 
     Filter by name and sort desc. For example:
 
-        vss-cli compute iso public ls -f name like,Cent% -s path asc
+        vss-cli compute iso public ls -f name=like,Cent% -s path asc
     """
     params = dict(expand=1, sort='name,asc')
     if all(filter_by):