+++ vss_cli/plugins/compute_plugins/os.py
@@ -29,11 +29,11 @@ def compute_os(ctx):
 def compute_os_ls(ctx: Configuration, filter_by, page, sort, show_all, count):
     """List requests based on:
 
-        Filter list in the following format <field_name> <operator>,<value>
+        Filter list in the following format <field_name>=<operator>,<value>
         where operator is eq, ne, lt, le, gt, ge, like, in.
         For example: status eq,PROCESSED
 
-            vss-cli compute os ls -f full_name like,CentOS%
+            vss-cli compute os ls -f full_name=like,CentOS%
 
         Sort list in the following format <field_name> <asc|desc>. For example: