+++ vss_cli/plugins/compute_plugins/template.py
@@ -32,11 +32,11 @@ def compute_template_ls(
 
         Filter and sort list by any attribute. For example:
 
-        vss-cli compute template ls -f name like,%vm-name% -f version like,%13
+        vss-cli compute template ls -f name=like,%vm-name% -f version=like,%13
 
         Simple name filtering:
 
-        vss-cli compute template ls -f name %vm-name% -s name desc
+        vss-cli compute template ls -f name=%vm-name% -s name desc
 
     """
     params = dict(expand=1, sort='name,asc')